Downgrade Mongoid to 2.8.1
Mongoid 3.0.x does not return MongoDB errors when
inserting a row breaking a uniqueness constraint.
This results in nasty behaviour in the model because
it acts as though it has been persisted when it
actually hasn't.

There doesn't seem to be an easy fix due to the way
Moped makes requests to MongoDB, so let's bite the
bullet and downgrade to Mongoid 2.x.

This involves modifications to config/mongoid.yml,
changes to how we override the IDs in both the
Organisation and Need models, and a change to the
validation of organisation IDs for a Need.

There's also a test for the Need ID uniqueness
behaviour to assert that an exception is raised.