List just the basic information about each need in result sets
So that we can include more fields, like the event
history, in single need responses, we should
return fewer items of information in the result
set responses.

This creates a new BasicNeedPresenter which returns
only the role, goal, benefit and organisations for
a need. This changes the NeedResultSetPresenter to
present needs using the basic presenter instead of
the full presenter.