Add a need decorator to return pairs of revisions
Add a decorator method to return revisions as pairs 
with their previous revision for comparison.

This didn't feel like it was a model concern, nor
did it fit cleanly to be a controller concern whilst
maintaining a single need parameter to the 
NeedPresenter.

This also includes a new FactoryGirl factory for
a need revision.