Add new field "applies_to_all_organisations"
This boolean field should be true if the need is
one which we would consider "generic".

The word "generic" doesn't really mean much, so
'applies to all organisations' seems a bit fairer.

There's a bit of hackery to get the boolean field
to default to false, even when it's set to nil.