Disallow organisation IDs for generic needs.
Since the intention is that the `applies_to_all_organisations` flag, if
set, overrides the `organisation_ids` field, it makes sense to be
explicit and not allow any IDs to be set.

We're only expecting to instantiate a small number of these needs, and
not allow extra ones to be set in Maslow itself, so we shouldn't need to
worry about someone trying to set this flag on a need that already has
organisations set.