Add an "in_scope" boolean field to Need
This adds a new boolean field to Need, "in_scope", which is used to
indicate whether a need is in scope for GOV.UK.

Right now, we only want to be able to mark a need as out of
scope, as there will be future work to add a workflow for deciding that
a need should be in scope on GOV.UK. To reflect this, the in_scope field
can only be nil or false at the present time.

The in_scope field is indexed in search so that it is returned as part
of the need list view.