Prevent concurrent Organisation imports
This rake task is run every hour on every machine that the app is deployed to.
It only needs to be run by one of them, but rather than make it fragile by
running on a specific server, we should try to run it on every server but use a
distributed lock to prevent unnecessary work.

This is more of an issue than it used to be because we are running the task
once per hour rather than once per day.