Paginate search results
Make search results paginatable. This also shortens and simplifies the code.

Previously it was silently ignoring pagination params if supplied.

Had to switch from `count` to Kaminari's `total_count` method because, while
either works with a Mongoid result set, `count` returned the size of the array
when called on an array.