Upgrade Mongoid
Upgrades Mongoid dependency to the latest in the `4.X` series. The `5.X`
series is reasonably new and we've yet to adopt it elsewhere.

Summary of the changes required to get the test suite back to green:

* Remove `attr_accessible` in favour of strong parameters
* Replace `key` usage with `_id` and `default` option
* Update `index` usage to new syntax
* Update `scope` and `default_scope` usage providing lambdas to wrap
 criteria
* Verify timestamps strings against the default Rails format. These are
 still parsable by Maslow as the equivalent `iso8601`
* For testing, create a `Need` with a given `need_id` as `_id` is now
 immutable after creation

The replacement of `update_attribute` with the equivalent `update!` was
due to an error being raised thanks to a frozen hash mutation happening
somewhere deep in the bowels of Mongoid.

Both @binaryberry and myself went down the rabbit hole tracing where
this was occuring and pinpointed it to a frozen `NeedStatus` attributes
instance. Finding the code that mutates this instance would require a
bit more time, so we've parked it and pencilled in to revisit soon.