Exclude some fields from export task
We do not want to add deprecated attributes into the Publishing-API. in_scope
and out_of_reason were being used by 11 records, and are no longer in use as
they've been replaced by NeedStatus. The currently_met and monthly attributes
are no longer in use either. We can revisit whether we keep that data when or
if we decommission the Need API.

Also ignoring content-id as it is not needed in the details. It is provided as
a base field.