Filter revisions to match Publishing API workflow
This maps need statuses to the relevant Publishing-API draft or published
states.

It ensures only the latest version of a proposed need becomes a draft in the
Publishing API. When several need_revisions have been marked as valid or valid
with conditions, it sets the state of the latest version to "published", and
the other ones as "superseded".