Export need_id for each need
Need Content Items require keeping a reference to need_ids, as they are still
used in Maslow, so this adds them to details. The need_ids array, which is in
the base schema, is used for tagging content items to needs, so not relevant
for storing the need_id of needs themselves.