Update to rails 5 compatible version of kaminari
Latest is 1.x but that introduces many changes we don't want to deal
with and v0.17 is enough for rails 5 compatibility.  This does introduce
a deprecation warning about mongoid functionality being extracted to a
separate gem in the next release of kaminari, but it's a one-time message
at startup so I think we can live with it for now.  It'll serve as a
pointer for future work when we do decide to upgrade to kaminari 1.x.