Stop configuring assets
We don't have any assets in this app because it's an API app.  We don't
include sprockets and in rails 5 this means we don't have the
`config.assets` object to configure.  Much like the default active_record
configuration options, we comment these assets ones out instead of
removing them entirely to make it easier to update in the future.