Avoid stack trace too deep while saving needs
After upgrading to rails 5 and mongoid 6 saving a need caused an infinite
callback loop. Saving a Need causes one of the after_create/after_update
callbacks that build a new NeedRevision to run.  Saving this object in
turn causes the need itself to re-save (perhaps as part of a touch
event?).  This then triggers another after_update callback to run and
build a new NeedRevision, which in turn ...

To avoid this we make sure we set inverse_of arguments so we are always
dealing with the same objects and then add guard clauses round the call
to create the NeedRevision in Need.  If we're alread inside a callback
we won't run it again.  Once we successfully create the NeedRevision we
discard the guard and fall out of the callback.