Use rails 5 version of strong_parameters
This means we make sure we use the strong parameters API where we can,
and also that we call `.to_h` on the outcome of using the API to convert
from the rails 5 parameter object to the hash we used to get.  We also
have to deal with differences for how empty params are parsed.  Before
rails 5 a query string like `?foo=&bar=baz` would be parsed into
`{ foo: nil, bar: 'baz' }`, but in rails 5 it is parsed into
`{ foo: '', bar: 'baz' }`.  Ultimately this means checking values with
`.present?` instead of relying on nil checks.