Turn off mongoid 6 validations on belongs_to associations
In Rails 5 belongs_to associations in activerecord have a default
validation to make sure the object at the other end of the association
is present.  Mongoid 6 also introduces this functionality for parity with
activerecord.  We didn't have this validation before, and it's not clear
to me if we should have it, but it's easy enough to disable the
functionality in an intializer.